dom/svg/test/test_onload2.xhtml

<html xmlns="http://www.w3.org/1999/xhtml">
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=1474311
-->
<head>
  <title>Test onload behaviour</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=1474311">Mozilla Bug 1474311</a>
<p id="display"></p>
<script class="testbody" type="text/javascript">
<![CDATA[

SimpleTest.waitForExplicitFinish();

function run() {

  let doc = document.implementation.createHTMLDocument('');

  doc.body.innerHTML = '<svg xmlns="http://www.w3.org/2000/svg" xmlns:xlink="http://www.w3.org/1999/xlink"><image id="image" xlink:href="data:image/gif;base64,R0lGODlhAQABAIAAAP///wAAACH5BAEAAAAALAAAAAABAAEAAAICRAEAOw==" height="200" width="200"/></svg>';

  let img = doc.body.firstChild.firstChild;

  document.getElementById('svg').appendChild(img);

  img.addEventListener('load', function () {
    ok(true, "onload method called");
    SimpleTest.finish();
  });

  img.addEventListener('error', function () {
    ok(false, "onerror method called");
  });
}

]]>
</script>
<div id="content">

  <svg id="svg" onload="run()" xmlns="http://www.w3.org/2000/svg" xmlns:xlink="http://www.w3.org/1999/xlink" width="100%" height="1">
  </svg>
</div>
<pre id="test">
</pre>
</body>
</html>
